lopper_tree: add alias lookup functionality
LopperTree now tracks node lookup by alias, if an alias node is in
the loaded data.

By passing the alias string to the alias_node() method, you'll get
the target LopperNode or None (depending if the alias is valid or
not).

Also note that this supports aliases that use labels in their paths.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>