xlnx: xppu/xmpu: add permitted apertures for boot protections
Due to a few platform specific dependencies at boot, there is a need
open up few protected regions for specific masters. Without this, the
boot will not work on the platform. These protected regions are
referred to as "RegNodes" (or Register Nodes).

Currently, there is no S-DT/YAML specification available to express
protections for RegNodes. Due to this, the output CDO directly needs to
be patched with hardcoded entries so the boot will function. These will
be removed in later when the gaps in the specification are resolved.

Signed-off-by: Izhar Shaikh <izhar.ameer.shaikh@xilinx.com>