tree: make label search exact by default
When a label lookup was executed, the label name was used
as the regex. This will of course match labels that contain
the target label as a substring.

Add a flag that is true by default and use it to bound the
label search with ^ and $, so we get an exact match

If something was relying on the old behaviour, they can pass
False to the lnodes() call to get the old matching.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>