lopper: add module and python string to debug options
When debugging, you often just want to see what the loaded tree looks
like, or perfom some simply queries before lops are executed.

To allow this, we add the ability to load a python file (where a
function of the same name as the file will be called), or take a
python string directly and execute them in the debug context.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>