yaml: add new configuration knob 'scalar_as_lists'
The way that yaml reads are encoded to a tree is different than the
way that fdt based reads are transformed.  This is due to the need
for json and other formatting .. the values are not lists, but are
instead single values.

This is unfortunately encoded into many handlers, so we can't do
a global conversion to LopperProp's where "value" is always a list
(like it is from libfdt).

So we can offer a configuration option that switched the behaviour
when a handler is expecting this format:

  --cfgval yaml.scalar_as_lists

And all scalar values will be encoded as lists into the LopperProperty

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>