lopper/yaml: add lopper processing for "<<+" extension marker
To allow enhanced inheritance and inclusion not possible with
standard yaml anchors and aliases, we introduce a lopper extension
marker of "<<+"

When this is found in the input yaml stream, lopper will perform
extra expansion and unwiding on aliases that are found in the
input stream.

These are still valid yaml keys, as such they will flow through
a standard yaml front end, but will remain in the output as-is
without additional expansion or processing.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>