lopper/lops: resolve tree when copying
If a tree has labels, they won't all be present when doing the
first pass of property resolution. To be sure that we've fully
resolved all properties we trigger a second full tree wide
resolve after the load.

We also use non-strict resolution on load, to ensure that no
records are dropped.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>