tree: fix tree export and reload
Some information was being dropped when an exported tree was
reloaded (labels, phandles, etc). On load, we now ensure that
non-strict resolutin is used, and we restore pclasses for
specially named properties.

We also enhance the phandle resolution to optionally return
context fields. This is useful when a phandle is part of a
set of fields, and information from other fields is needed
to process a phandle.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>