lopper/format: allow tabs to be used when writing dts
To make comparisions easier, it is sometimes desireable to write dts
indentation with tabs versus spaces.

We make our output more flexible, so it can use either.

We also use the existing cfgval option to make the choice selectable
on the command line:

  --cfgval dts.tabs=True

To output using tabs

The default stays as spaces.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>