fdt: make node prefix lookups more robust
libfdt returns matches based on partial paths. i.e. if a tree
has /timer@2a810000 and /timer, and you search for /timer, you
will get the node number for /timer@2a810000 .. this causes problems.
so we double check the name, and if it isn't as expected, we do a
slower secondary search to find the proper node.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>