assists: create generic tree extraction and xen-processing assists
To support the extraction of a tree and subsequent xen processing
of an extracted tree, we introduce a generic extract assist and a
xen specific assist.

The generic extract assist takes a target node, and will extract
it from the system device tree, placing it in a new tree, contained
within an "extracted" top level device tree node.

The generic extract will then chase phandle references of that
node, and pull them into the same extracted tree structure, doing
appropriate fixups to paths and properties.

The generic extract can also take a list of nodes and properties
to drop during the extraction process. While these sorts of
exclusions can also be done via lop files, it is convienient to
allow it in the assist, since that makes pipeline processing
easier.

Once complete, the extract assist places the tree within the
System Device Tree 'subtrees' dictionary, under the key
"extracted". This allows subsequent assists to continue processing.

The extract-xen assist, renames elements of the extracted tree
into names, properties, etc, that Xen expects. It also adds
properties and other components that are required for passthrough
functionality.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>