lopper/tree: handle invalid last phandle
When we are formatting a list of phandles, lopper looks ahead and uses
a , or ; depending on where we are in the list.

If the last phandle is invalid, and is dropped, we can end up with an
incorrectly terminated list (with a ,).

So we detect this situation, and overwrite our , with a ; in that scenario.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>