tree: make label generation more robust
Lopper was generating duplicate labels, which can cause all sorts
of issues with the output device tree, when two nodes had the
same name.

It was also generating labels for every node, regardless of whether
or not it was a phandle target.

We can improve both things:

  - generate a label when a phandle reference is processed
  - make sure the label is globally unqiue, and update the LopperTree
    to track the node by that label

We leave the start of a flag to control whether or not labels are
always generated, but always set it to False for now.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>