tree: ensure that phandle is initialized before export
The export of nodes was changed to include phandles a few weeks
ago. Which works fine when a node has a phandle, but if the node
doesn't have a phandle, the old default of '0' is now '-1'.

'0' had significance when processing references, '-1' doesn't.

So we check the phandle before export, and if it is -1, we set it
to 0.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>