tree: handle variable sized phandle properties
The lopper tree routines were capabable of using a #cells valuey
to determine where phandles were located within a property, but
it also incorrectly assumed that the value was consistent, and
phandle locations could be calculated for every set of fields
within a given property.

This isn't true, and resulted in some values being dropped as
properties were being exported.

We fix this by introducing a new routine phandle_map() that is
capabable of returning the location of all phandles and non
phandle fields within a property. We adapt our routines to use
this base function when writing a property to the various
output formats.

This is a fundamental change, and we should expect some
fallout in corner cases.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>