fdt: fix dtb handling of common prefix nodes
libfdt matches "<name>@<address>" to nodes of just "<name>", this
is seen with a node "timer" and existing nodes of "timer@beef".

When this happens, the existing routines don't detect it, and will
eventually add the properties of the conflicting node to node 0
(i.e. the root node).

We can create a temporary node name with a "lopper" prefix, to
get it added, and then rename/delete it later to avoid the
issue (there's no way to fix it in libfdt).

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>