lops: create phandles for modify operation references
When a modify lop involves a node reference, we need to ensure that
the node has a valid phandle. The node may have been newly added,
never referenced, or otherwise modified to no longer have a valid
phandle.

We know that the modify operation is creating a reference, so
so generating and assigning a phandle is in order.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>