specification: convert to Sphinx
Convert the core system devicetree specification to the
restructuredtext file format (.rst) and Sphinx documentation system.

This is the same documentation system used by both the (regular)
devicetree specification and the Linux documentation.

Using this format over plain markdown has many advantages. A few are:

- auto-numbered sections
- HTML output
- PDF output (via LaTeX)
- the ability to easily reference other sections
- syntax highlight for DTS fragments

Include a README.md for how to build the source. To avoid bitrot and
skew, delete the .md documents which were converted to the new format.

I took care to make the minimum number of formatting changes possible
during this conversion, avoiding typo fixes, unnecessary formatting
changes, etc. This is meant as a starting point for further
refinements where each .rst page should be easy to compare to its
starting point.

I left a few documents out of the conversion when it wasn't clear that
they were part of the 'core' system DT spec. If that's wrong and they
are meant to be part of the spec somehow, that's easy enough to fix as
well.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>