specification: fix incorrect usage of unit addresses
From the devicetree specification, section 2.2.2 Node Names:

    The unit-address must match the first address specified in the reg
    property of the node. If the node has no reg property, the
    @unit-address must be omitted and the node-name alone
    differentiates the node from other nodes at the same level in the
    tree.

From this we can also infer a derived requirement that any node with a
unit address must have a parent node with #address-cells set, and the
reg property must be consistent with it.

Fix unit addresses appearing in specification/source and
specification/system-device-tree.dts to meet these requirements. Any
remaining unit addresses in those files -- except for those in
chapter3-secure-addresses.rst -- should be following these rules;
those that don't are just bugs that I missed.

chapter3-secure-addresses.txt is going away soon, so I'm just going to
leave it as is.

I didn't look anywhere else in the tree as I am focusing on the core
specification.

Fixes: https://github.com/devicetree-org/lopper/issues/128
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>