specification: remove secure-bus and secure-reg
The secure-bus and secure-reg properties were originally added as a
result of discussions how to represent address maps for Arm v8-M CPUs
with TrustZone support. In this case, the same IP blocks often
effectively have different addresses depending on the security state
of the CPU at the time of access.

They haven't been useful in practice and are not used anywhere in the
lopper repository. Remove mentions to them (and renumber the documents
as needed).

The manual renumbering is only required because the file names contain
their chapter numbers. This was done for consistency with the DT spec
source. We can remove the chapter numbers from file names for now if
we continue to insert and remove chapters in the middle of the spec
document.

Fixes: https://github.com/devicetree-org/lopper/issues/130
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>