specification: move bus firewalls chapter out of sphinx
After further discussion with Stefano Stabellini, the 'core' of the
system DT specification at present is currently about cpu clusters,
indirect buses, and execution domains.

Restore the bus firewalls markdown file, removing the .rst version, to
allow us to concentrate on the core without removing this content. We
can always restore the chapter in the spec later if need be.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>