lopper: add json input/output file format
This is the introduction of json read/write support for Lopper.
We are heavily re-using the Yaml infrastructure to support this
new format.

This is not heavily tested, but we can now read a json file to
a tree, and write a tree back to a json file.

The naming of the "yaml" module will change to represent the
new flexibility (in subsequent commits)

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>