specification: rework intro and add glossary/bibliography
Rework the introductory text to be more formal and correspond roughly
to the structure of the base devicetree specification. As part of
making this more formal, add a glossary and a bibliography to avoid
undefined terms or references to other pieces of software without
context.

Note in particular that we are moving from "Device Tree" to
"devicetree"; this is the preferred way to write it.

Remove some extended descriptions of details; these detract from the
overview nature of an introduction and should be in the individual
chapters comprising the core of the specification.

Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>