specification: migrate remaining .md documents to Sphinx
The initial chapters of the specification have stabilized to a
significant extent compared to the remaining .md files in the
specification directory.

However, it would be good to have the entire spec under a single
source tree, without maintaining two separate document formats like we
have now, which could be confusing.

Moving everything to Sphinx might create the perception that
everything in Sphinx is equally stable, however. That isn't true. To
solve the confusion without potentially creating more, import the
remaining .md sections of the specification into Sphinx, but clearly
mark them as "under discussion".

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>