lopper/tree: create address() lookup on a node
As a building block to returning a complete memory map, or looking
up a device by memory address, this commit introduces an address()
routine to LopperNodes.

When called this method returns the mapping of the node in the tree.

This is completed by recursively walking the Nodes hierarchy and
interpreting any unit address or ranges properties that are found.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>