lopper: handle json encoded strings in DTC
Complex formats like json have structures that are not able to
be expanded carried as json encoded strings.

If we write those strings to a .dts file dtc won't be able to
parse, since nested quotes are invalid.

To fix this, we escape any nested quotes, and then do a test json
decode when writing to complex formats like yaml. this allows
output formats of both dts and json/yaml to be supported.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>