lopper: add "cond" property to lops
"cond" specifies a target lop phandle. If the result of that lop
is "True", then the specifying lop will be executed. If the result
of the target lop is False, then the lop will not be exected. This
is typically used to trigger execution of lops based on selected
nodes, or a property found in a tree.

In the following example, either lop_1 or lop_1_1 are valid targets
of the "cond" property of lop_1_1_1. The result would be the same if
either was specified (and both are provided to show a long and short
form of the operation). If a matching compatible node is found in the
tree, then lop_1_1_1 will be executed.

            lop_1: lop_1 {
                  compatible = "system-device-tree-v1,lop,select-v1";
                  // clear any old selections
                  select_1;
                  select_2 = "/:compatible:.*xlnx,versal-vc-p-a2197-00-revA.*";
            };
            lop_1_1: lop_1_1 {
                  compatible = "system-device-tree-v1,lop,code-v1";
                  code = "
                          if __selected__:
                              print( 'Compatible dts (type1) found: %s' % node )

                          if __selected__:
                              return True
                          else:
                              return False
                      ";
            };
            lop_1_1_1 {
                  compatible = "system-device-tree-v1,lop,code-v1";
                  cond = <&lop_1>;
                  code = "
                         print( 'Conditional Code is running!' )
                         ";
             };

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>