yaml: override default anytree iterator generator
When exporting to a dictionary the nodes in the AnyTree are
walked, and non-internal nodes + properties are exported to
a dictionary structure.

By default Anytree uses an iterator generator that produces
a list of tuples (which is standard). When the list of tuples
is used to initalize a dictionary, the keys are sorted
alphabetically.

This sorting is not an issue in some use cases, but in ours,
it means that properties of a tree (dts, yaml, json) will be
sorted instead of having their source order maintained.

Again, yaml/json don't demand that ordering is maintained,
and the result is the same when interpreted .. but a user
or someone editing the output will be surprised by this change.

We make two changes to maintain the order:

  - Use an OrderedDictionary class in all cases
  - Override the iterator to generate a full list, versus
    yeilding the key value pairs.

We also override the output of yaml tags, since switching
to an OrderedDict triggers !!odict to be added to nodes.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>