assists/isospec: introduce reference domains.yaml generation
The yaml convenience / simplified representation of the
system device tree domain specification has been documented
for quite some time.

This assist shows how a device specific, non standard specification
of isolation information can be used to generate the equivalent
system device tree domain specification in yaml.

This is useful to take non-open formats and convert them to
a human readable and editable format, which is part of the system
device tree specification.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>