specification: zero cpu-mask is "all the CPUs"
Extend the execution domain binding to say that a cpu-mask cell value
of 0 means that the domain runs on all of the CPUs in the cluster.

This seems like a common case, and it is a useful shorthand which may
help users, who could define a self-documenting macro like
"DOMAIN_CPU_MASK_ALL" or so for this.

Fixes: #166
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>