specification: domains: clarify start and size fields
The 'start' and 'size' fields in the 'memory' and 'sram' properties of
the execution domain binding could use a bit of clarification. In
particular, it's not fully specified where the #address-cells and
the #size-cells properties that determine the number of cells used
for these fields come from. Spell it out to be explicit.

Fixes: #162
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>