specification: domains: fix access property semantics
Following some discussion, it became clear that the device access
granted by a domain's 'access' property is not intended to be
exclusive. However, the specification currently states that it is.

This is unintended and precludes use cases where two domains need to
share a resource. Fix it by saying access does not imply exclusive
access, as intended but currently not as specified.

Fixes: #164
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>