specification: domains: make id sometimes optional
Previous use cases for execution domains generally required an id
property. However, as the specification is applied in more contexts,
it's becoming clear that it may not always be needed. In particular,
the Zephyr RTOS is not expected to have initial use cases for this
property (though it may in the future start using it).

To avoid forcing these users to maintain unique id properties that are
not needed, refine the binding so that 'id' is only needed in some
cases.

Fixes: #163
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>