yaml: fix output to not be list based
When the switch to OrderedDicts was made, the export of these
dictionaries results in yaml output that is list versus mapping
based.

It is easier to read and process mapping based yaml (at least
for lopper purposes), so we flatten the OrderedDicts to dicts
and use ruaeml's rountrip mode to preserve attribute order when
writing files.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>