lopper: merge inputs with domains to main device tree
Existing assists expect that domain specifications (whether yaml
or dts) are merged into the main device tree.

commit b548dae3a652d3e3 [lopper: allow input files to be stored for future use]
was tagging them as "support files", and hence they were not merged.

The logic for detecting domain files was only triggering on
subsystems. We add a check for a domain-v1 compatible node, and if
found, we allow it to be merged into the system device tree.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>