bmconfig: Change in the logic to get the _g.c name for the drivers
Currently we are relying on the Config sturcture name to get the _g.c file name.
Instead we can rely on the _g.c file name already present in the driver folder
and create the _g.c file with the same name as the output of this assist.
This solves the issue with devcfg driver where the Config name is DCfg but the
driver name and the file name starts with devcfg

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>