isospec: format memory ranges in hex
Rather than use numbers (which are hex), we can format the memory
ranges as hex. This outputs them to yaml as strings/literals, but
we can easily convert them to numbers if manipulations are required.

Having them as hex, makes comparisons to the device tree easier, as
those addresses are all hex format.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>