isospec: handle subdomains of default subystem
The previous multi-domain support expected that only subsystems
existed and would be peers in the tree.

As it turns out, an isolation more typically has a subsystem, within
domains defined as children in that subystem.

We re-work the compatibilty if our domains (using domains versus
subsystem), and factor the processing out so we can detected and
nest domains.

We additionally add a check for the requested flag, and if not
present, we don't include a device/label.

Finally, we keep the device tree label in the yaml as a separate
property, so that information won't be lost on device lookup by
address.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>