lopper:assists:baremetalconfig : Adding support to read a 64 bit property from a node on demand
By default, reg and ranges are the only 64 address properties which have specific handling in the
baremetal config assist. For PCIE, there is a requirement to read two 32 bit cells of hex values
as a 64 bit address and not a list of two different hex values. This value can be different than
the existing reg property. Thus, an additional usage of padding is being introduced here to
support treating multiple cells of hex values as one big address on demand.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>