lopper: assists: baremetal_xparameters_xlnx: Ignore Subnodes
The subnodes within a device (where they exist) are not
necessary to be included in xparamaters.h, the base paramaters
are sufficeint.

This patch excudes the submodes by using a break on finding
a dict withing the node being parsed.

"nosub" is also a command that should not be included,
a continue is used for this instance.

Signed-off-by: conallogriofaamd <conall.ogriofa@amd.com>