lopper:assists:baremetal_config : Add flexibility for padding number
7f0464d75 commit added the support to get a 64 bit property in the driver
config structure. That change expected the padding number to be same as that
of number of cells present in that property. This adds a flexibility to the
existing case where maximum possible number can be added in driver yamls
and the property can have number of cells less than or equal to that number.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>