assists:baremetal_xparameters_xlnx: Fix for the wrong reuse of variable "index"
c4c0b126fc8f5b commit added support to get a 64 bit property in xparameters.h. While
adding the support, the "index" variable was wrongly used as a loop variable. The
"index" variable was already declared before as a top level loop variable and this
led to a wrong canonical entry in xparameters file. Renaming this index loop variable
to fix this issue.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>