lopper:assists:baremetal_xparameters_xlnx: Removing the BOARD dependency on compatible string
Currently, there are two mechanisms to define XPS_BOARD_<board_name> in xparameters.
One is to define a compatible string in the root node and other one is to define a
new property "BOARD" in the root node. Both of these are derived in the same way at
the time of sdt generation and are redundant in nature. Therefore, removing the
dependency of mentioned macro's definition on the compatible string. Compatible
property in root node was also creating overwriting issue among the multiple
compatible strings available within the board specific dts files.Thus, using BOARD
for this makes more sense.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>