lopper:assists:baremetallinker_xlnx Correcting logic to get the stack and heap data from yaml
When "if" condition is used without any right hand comparison, logic returning 0 is treated
as False in python. In case when 0 heap size has to be put in yaml, the existing logic is
treating it as False statement and moving on. Therefore, correcting this logic to get the
correct heap and stack size in linker script under all possible scenarios.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>