tree: enhance phandle handling for variable sized fields
Previous fixes to allow a phandle description to work with
variable sized fields inadvertently degraded the handling
of descriptions that used constant spacing.

We fix this by ensuring that the phandle checking routines
note if variable or constant fields are being used and
increment the checking indexes appropriately.

Also, we tweak our phandle generation routines to create
a phandle property, versus only assigning it to the node.
This allows the phandle to be explicitly seen, without
needing to look at a fdt.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>