assists/domain: respect -t flag to lopper
When lopper is run with a target domain (-t) the domain_access
assist should check for that value and locate the node.

If the node is not found, we error.

Note: this only applies to command line runs. When the assist
is called from a lop file, the target_node will already be
the desired domain.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>