yaml: resolve json properties after setting pclass
To save processing times, we move the json pclass set to
after the property is created (since the property will
clobber it to "" during init), and we don't resolve
the property until it is set.

We have a short circuit in the property resolution to
not iterate json properties, so this saves overall
processing time.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>