tree: improve phandle map variable field handling
Not all looked up values behave the same way, #address-cells
or #interrupt-cells behave differently than #clock-cells. Their
behaviour is defined in the device tree bindings.

To allow for variable sized field processing when doing phandle
lookups and symbolic replacement, without needing to hard code
property names, we extend the lopper phandle specification to
include an expression as well as a lookup value.

This allows us to adjust the found value to match the correct
amount of fields to skip in a data driven manner.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>