tree: count indexes for dropped records
When we detect an invalid phandle and we are in strict mode,
the record containing the phandle is dropped.

This means that the processing for the index is skipped, which
includes incrementing the index to the next value. As a result,
we are forever offset and creating an output that doesn't
match the input.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>