openamp: xlnx: Add new module flags for role, host and remote
Enable usage of module using explicit key value pairs as command line
arguments

The arguments required now are:

--openamp_role=[host|remote]

Must be 'host' or 'remote'. This denotes the role which the output header file
will target.

--openamp_host and --openamp_remote

These each must correspond to the relevant cores in a OpenAMP Channel Pair.
On ZynqMP if a YAML has A53-0 as host and RPU0 as remote a valid command line
set would be:

--openamp_role=host --openamp_host=a53_0 --openamp_remote=r5_0

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>