openamp: xlnx: Enable Linux generated Device Tree to work for OpenAMP use cases
Update pruning in gen_domain_dts so that:
- TCM banks remain. Need these for OpenAMP ELF Loading
- Remove extra IPI nodes. These conflict if other domains want to use IPI too.

Add lop file to invoke gen_domain_dts. Make sure it is present in lop-load.dts
so that the assist gen_domain_dts can be invoked as a lop file. This enables
user to control the order of assist invocations upon the input SDT.

Sample usage:
./lopper.py -f -v --enhanced --permissive     \
 -i <OpenAMP YAML>                            \
 -i lopper/lops//lop-load.dts                 \
 -i lopper/lops//lop-xlate-yaml.dts           \
 -i lopper/lops/lop-openamp-invoke.dts        \
 -i lopper/lops/lop-gen_domain_dts-invoke.dts \
 -i lopper/lops/lop-a72-imux.dts              \
 system-top.dts out.dtb

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>