isospec: ensure that memory and cpus are always lists
If a single cpu or memory is part of a domain, the yaml encoding
is a node. If there are multiple, then it is a list of maps.

The processing of the two types of yaml are quite different.

To simplify assists down the pipeline, we add an empty dictionary
when only a single memory or cpu is detected. This normalizes
the yaml output and makes the assists simpler.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>