asists/subsystem: make processing more robust
The processing of both cpus and memory can contain entry
dictionaries to signifgy the end of the list. We check for
this to avoid errors when processing.

The parsing of memory size can be more robust by trying
a string, a hex and then humanfriendly parsing. Only when
all three fail, should we assign a default.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>