assists/isospec: log new spec type ss_management
By default any unrecognized isolation type is an error. This
ensures that additions to the input files are not ignored and
the output domain information is incorrect.

A new type ss_management can be safely ignored, as it is not
for software. We log it when detected, but otherwise do nothing
with it.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>