assists/domain_access: fix memory node -> address-map updating
The current updating of the system device tree's memory node
wasn't taking into account all of the potential memory regions
within a domain.

With this update, we now search for memory ranges that need to
be updated by checking the memory entries in a domain against
the addresses found in the memory nodes of a device tree.

If there's an address match, and the size differs, we update the
system device tree memory entry to match (the 'reg' property).

If a memory node does get updated, we then search the tree for
any address-map entries that reference the memory node, and
check their address mappings. If the addresses match, we update
the address-map entry to be the same.

This allows the domains information to keep all the memory
references in sync and matching the domain specification.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>