assists/domain_access: update start and size of memory
The previous implementation could modify the start address of
a memory node (and associated address-map entries). But a
domain can specify both a start and a size, both can differ.

Updating the logic go handle both.

We also change the way that address-maps are updated. When
any part of a memory node is modified, we expect it to be
fully specified. So when we start updating the address-map,
we drop all the old mappings for a given memory node and
write all new entries for it (from the values pulled from
the domain specification)

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>