lopper: venv: add requirements.txt and venv usage in README
Using python3's venv is very powerful and makes doing lopper development
and usage easier.

More on it: https://docs.python.org/3/library/venv.html

Update the README to add usage instructions about this and a
requirements.txt which makes installing the lopper python dependencies a
breeze using pip.

Been using this flow for a while, it'd be a useful addition to lopper
mainline so others can also benefit from it.

While at it, fix some minor typos in the README as well.

Signed-off-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>