tree: return a list of nodes matching an address
In some device trees, child nodes with @unitaddresses which
are indexes will be returned at the same mapping as their
parent node when ranges is "ranges;".

To avoid having the child nodes clobber the parent node
(Which is what we want), we append a matching node when
found, if one is already present.

We adapt the isospec assists to accept a list of nodes, versus
a single match.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>