assists:baremetallinker: Correct the logic to set the start address of linker section for microblaze
538f6b5c7171 commit (assists: baremetallinker: Correct the start address for
text and data sections for microblaze) added the logic to start the BRAM or the
DDR addresses from 0x50 for all the Microblaze designs. While doing so, the logic
to check for the start address condition was missed. If the memory is not falling
under first 0x50 bytes which is reserved for reset vectors, there is no need of
this manipulation. Fix the same.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>