lopper: Update check to adjust DDR start and size address
Update check to adjust the start and size addresses only for the DDR
regions starting from 0x0.

previously we are adjusting the DDR start and size for both:
- DDR start address is 0x0
- DDR start address is 0x100000 (1MB)

the allocation of the pre-reserved 1MB memory space, the memory access
operations targeting the final 1MB of DDR  are experiencing
functionality issues.

Signed-off-by: M Lakshmaiah <lakshmaiah.m@amd.com>