lopper:assists:baremetallinker: Do not consider linear SPIs for memory tests
Recently we have started showing linear SPIs in the usable linker sections. Though
it can be used as a linker section, our existing memory tests src is not capable
of testing it and the test hangs.

Add a check that ignores the linear SPIs during creation of the list of available
Memory Controllers in the CmakeLists. This way, the linear SPIs will not be added
in the memory test config structure and avoid the hang.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>